USER INTERFACE AND STEP SIZE UPDATE!
Found out that 43 steps is approximately 1mm so that means 43000 steps is a 1 meter tow. This is a lot of steps and this means that a simulation of 1 tow takes a while to run. For just being interested in gaps and overlaps, make the steps smaller (smaller tow lengths), but generate more tows.